move readControlFile to testUtils
cleanup
